+++ Load_Planning_Project/employees/urls.py
@@ -10,4 +10,5 @@ urlpatterns = [
     path('<str:table_name>/<int:object_id>/edit', views.edit_record, name='edit_record'),
     path('<str:table_name>/<int:object_id>/del', views.del_record, name='del_record'),
     path('<str:table_name>/export', views.export_csv, name='export_csv'),
+    path('<str:table_name>/import', views.import_csv, name='import_csv'),
 ]