+++ Changelog.md
@@ -1,3 +1,17 @@
+###17/11/2020
+
+- Added Employees table
+- Some minor customization of displayed tables
+- Inserted UNIQUE property for e_mail column in Employees table
+- CSV Export option
+
+To be done:
+
+- Validation of fields/moving some parts of code to models if needed
+- Messages about statuses - edited properly and such
+- Validation of UNIQUE columns -> form should pop out message about field (it does not now)
+- CSV Import option
+
 ###16/11/2020
 
 - Building basic views and forms for show/create new/edit/delete functions